refactor: extract result cli module
